move and rename user menu and clock
