+++ centreon/www/front_src/src/Header/UserMenu/SwitchThemeMode/index.tsx
