+++ centreon/www/front_src/src/Header/UserMenu/SwitchThemeMode/useSwitchThemeMode.tsx
